Bug fix in Spinbox.
Use of keyboard arrows was triggering float conversion errors for
Locale's configuration where comma is used instead of dots.
The solution delegates to the parent widget the total handling of float
conversions and employ the property setKeyboardTracking to get the
appropriate behaviour when using the text editor of the Spinbox.